requirements.txt
opecv-python
numpy
tensorflow
flask
matplotlib
pillow
cython